+++ MiniApp/Classes/core/Models/MiniAppReward.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+/// A MiniAppReward is the structure that is returned to the Mini App when a rewarded ad is displayed to an user
+/// The host app developer can provide a reward type and the amount earned for watching the ad
+/// - Tag: MiniAppReward
+public struct MiniAppReward: Codable {
+	let type: String
+	let amount: Int
+
+	public init(type: String, amount: Int) {
+		self.type = type
+		self.amount = amount
+	}
+}