+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## CHANGELOG
 
+### 2.x.x
+
+**SDK**
+
+- **Feature** Made MiniApp SDK a static framework
+- **Feature** Added ads for Mini Apps host
+- **Feature** Added Google mobile ads integration in a `MiniApps/Admob` pod subspec
+
 ### 2.8.0 (2021-02-01)
 
 **SDK**